+++ Samples/iOS-Swift/iOS-Swift/ViewController.swift
@@ -13,6 +13,14 @@ class ViewController: UIViewController {
             let user = Sentry.User(userId: "1")
             user.email = "tony@example.com"
             scope.setUser(user)
+            
+            if let path = Bundle.main.path(forResource: "Tongariro", ofType: "jpg") {
+                scope.add(Attachment(path: path, filename: "Tongariro.jpg", contentType: "image/jpeg"))
+            }
+            if let data = "hello".data(using: .utf8) {
+                scope.add(Attachment(data: data, filename: "log.txt"))
+            }
+            
         }
         // Also works
         let user = Sentry.User(userId: "1")