+++ Samples/tvOS-Swift/tvOS-Swift/AppDelegate.swift
@@ -16,6 +16,16 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             options.attachStacktrace = true
             options.sessionTrackingIntervalMillis = 5_000
         }
+        
+        SentrySDK.configureScope { scope in
+            if let path = Bundle.main.path(forResource: "Tongariro", ofType: "jpg") {
+                scope.add(Attachment(path: path, filename: "Tongariro.jpg", contentType: "image/jpeg"))
+            }
+            
+            if let data = "hello".data(using: .utf8) {
+                scope.add(Attachment(data: data, filename: "log.txt"))
+            }
+        }
 
         // Create the SwiftUI view that provides the window contents.
         let contentView = ContentView()