+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+- feat: Add basic support for attachments #875
+
 ## 6.0.12
 
 - fix: Crash in SentrySession.serialize #870